pyproject.toml
[project]
name = "sheep-seg"
version = "0.1.0"
description = "Privacy-first computer vision for animal welfare on small farms"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "pillow>=10.0",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "torch>=2.1",
    "torchvision>=0.16",
    "anthropic>=0.40",
    "pillow-heif>=0.18",
    "transformers>=4.40",
    "open3d>=0.19.0",
    "rembg[cpu]>=2.0.75",
    "trimesh>=4.11.5",
    "einops>=0.8.2",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[project.scripts]
sheep-seg = "src.cli:main"
